+++ packages/widgetbook/lib/src/addons/device_frame_addon/device_frame_addon.dart
@@ -2,7 +2,6 @@ import 'package:device_frame_plus/device_frame_plus.dart';
 import 'package:flutter/material.dart';
 
 import '../../fields/fields.dart';
-import '../../inherited_widgetbook_theme.dart';
 import '../common/common.dart';
 import 'device_frame_setting.dart';
 import 'none_device.dart';
@@ -89,16 +88,11 @@ class DeviceFrameAddon extends WidgetbookAddon<DeviceFrameSetting> {
           // workbench and not just the device frame.
           screen: Navigator(
             onGenerateRoute: (_) => PageRouteBuilder(
-              pageBuilder: (context, _, __) => ColoredBox(
-                color: InheritedWidgetbookTheme.maybeOf(context)
-                        ?.scaffoldBackgroundColor ??
-                    Colors.transparent,
-                child: setting.hasFrame
-                    ? child
-                    : SafeArea(
-                        child: child,
-                      ),
-              ),
+              pageBuilder: (context, _, __) => setting.hasFrame
+                  ? child
+                  : SafeArea(
+                      child: child,
+                    ),
             ),
           ),
         ),